Updated card identification procedures
Previously, I incorrectly treated the ATS as the historical bytes. As such, I was improperly checking whether there were any such bytes present: checking the ATS size instead of the data within the ATS. I have introduced a function to get the historical bytes correctly to address this issue, which will stop the mis-identification of DESFires as non-MIFARE devices, as well as other non-tested mis-labellings that may occur. I also corrected the functionality of the GetVersion command, which didn't work as intended due to mis-handling of the command.