Removed the get_target_version function
As the function was thinned down in its fix, its own function is no longer necessary. Additionally, the function wasn't returning the version anyways (misnomer). As far as the command's use in the tool up to this point, I simply need to check if the execution was successful, which I can do with the value returned by the transmission function.